Support for blocks and template inheritance
Improved template search mechanism. Templates are now searched for in
the current working directory, in /etc/xtables-monitor and on the whole
filesystem (to be able to pass an absolute path to the executable).

Improved template rendering. Generated rulesets are now printed line by
line instead of statement by statement.

Changed the examples to use blocks and inheritance. This makes the
examples much simpler and smaller.